+++ MissingEnvVarsError.js
@@ -1,25 +1,23 @@
 'use strict';
 
-var util = require('util');
+const util = require('util');
 
-function MissingEnvVarsError (allowEmptyValues, dotenvFilename, sampleFilename, missingVars, error) {
+function MissingEnvVarsError (allowEmptyValues, dotenvFilename, exampleFilename, missingVars, error) {
+    const errorMessage = `The following variables were defined in ${exampleFilename} but are not present in the environment:\n  ${missingVars.join(', ')}
+Make sure to add them to ${dotenvFilename} or directly to the environment.`;
+    const allowEmptyValuesMessage = !allowEmptyValues ? `If you expect any of these variables to be empty, you can use the allowEmptyValues option:
+require('dotenv-safe').config({
+  allowEmptyValues: true
+});` : '';
+    const envErrorMessage = error ? `Also, the following error was thrown when trying to read variables from  ${dotenvFilename}:\n${error.message}`
+                                  : '';
     Error.call(this);
     this.name = this.constructor.name;
     this.missing = missingVars;
-    this.sample = sampleFilename;
-    this.message = 'The following variables are defined in ' + sampleFilename + ' but are not defined in the environment: ' + missingVars.join(', ') + '.\n' +
-        'Make sure to add them to ' + dotenvFilename + ' or directly to the environment.';
-    if (!allowEmptyValues) {
-        var emptyValuesMessage = 'If you expect any of these missing variables to be empty, you can use the allowEmptyValues option: \n\n' +
-            "require('dotenv-safe').load({\n" +
-            '  allowEmptyValues: true\n' +
-            '});';
-        this.message += '\n' + emptyValuesMessage;
-    }
-    if (error) {
-        var errorMessage = 'Also, the following error was thrown when trying to read variables from ' + dotenvFilename + ': ' + error.message;
-        this.message += '\n' + errorMessage;
-    }
+    this.example = this.sample = exampleFilename;
+    this.message = [errorMessage, allowEmptyValuesMessage, envErrorMessage]
+        .filter(Boolean)
+        .join('\n\n');
     Error.captureStackTrace(this, this.constructor);
 }