+++ CHANGELOG.markdown
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [5.0.0] - 2018-02-19
+### Added
+- `example` option as an alias of `sample`
+- Testing against Node v4, v8 and v9
+
+### Changed
+- *Breaking*: Updated to dotenv 5.0.0. See [dotenv's CHANGELOG](https://github.com/motdotla/dotenv/blob/master/CHANGELOG.md#500---2018-01-29) for more details
+- Slightly reworded error messages
+- Updated code to use ES6 syntax features
+
+### Removed
+- Testing against Node v7
+
 ## [4.0.4] - 2017-04-28
 ### Fixed
 - Fix installation instructions for yarn. Released only to update README on npmjs.com.